added actuator, added provider side support for client credit line

+++ spring-cloud-contracts-provider/src/main/java/com/vgrazi/presentations/springcloudcontractsprovider/gateway/CreditIncreaseRequest.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 public class CreditIncreaseRequest {
     @Getter private double currentCreditLine;
     @Getter private double increaseAmount;
-    @Getter private double clientId;
+    @Getter private int clientId;
 }
